Implemented options page
Which shows each letter along with the cube faces it is near,
the current colour of each letter (in hex), and a sequence of
pairings of the letter with every other letter (to check that
the letter is distinguishable from the others).

More importantly the page allows you to alter the colours of the letters
and get immediate feedback on how that looks wrt the other letters.

A code snippet is displayed at the bottom with the current mapping